+++ docs/changes.rst
@@ -41,6 +41,10 @@ Released: not yet
 
 **Enhancements:**
 
+* Added an easy way to print debug information for inclusion into issues, via
+  `python -m zhmcclient.debuginfo`.
+  (See issue #640)
+
 **Known issues:**
 
 * See `list of open issues`_.